requirements.txt
# GazeAway - Eye Strain Reminder Application
# Dependencies for the GazeAway project

# System tray functionality
pystray>=0.19.4

# Image processing for system tray icon
Pillow>=9.0.0

# Windows API access for monitor detection and registry
pywin32>=305

# Build tool for creating executable
PyInstaller>=5.0.0

# Standard library modules (included with Python, listed for reference)
# - threading
# - time
# - os
# - json
# - sys
# - winsound
# - datetime
# - winreg
# - pathlib
# - wave
# - struct
# - math
# - shutil
